requirements.txt
Django==2.1.2
djangorestframework==3.9.0
#npm
#corsheaders